move typescript loader to own package
