+++ packages/typescript-loader/package.json
@@ -1,9 +1,15 @@
 {
-  "name": "cosmiconfig",
+  "name": "@cosmiconfig/typescript-loader",
   "version": "9.0.0",
-  "description": "Find and load configuration from a package.json property, rc file, TypeScript module, and more!",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    },
+    "./package.json": "./package.json"
+  },
   "files": [
     "dist",
     "!dist/**/*.map"
@@ -16,7 +22,6 @@
     "dev": "npm run build:tsc -- --watch",
     "lint": "eslint --ext .js,.ts .",
     "lint:fix": "eslint --ext .js,.ts . --fix",
-    "lint:md": "remark-preset-davidtheclark",
     "test": "vitest run --coverage",
     "test:watch": "vitest"
   },
@@ -42,6 +47,10 @@
     "url": "https://github.com/cosmiconfig/cosmiconfig/issues"
   },
   "homepage": "https://github.com/cosmiconfig/cosmiconfig#readme",
+  "dependencies": {
+    "@cosmiconfig/base": "9.0.0",
+    "@cosmiconfig/types": "9.0.0"
+  },
   "peerDependencies": {
     "typescript": ">=4.9.5"
   },
@@ -49,18 +58,6 @@
     "typescript": {
       "optional": true
     }
-  },
-  "dependencies": {
-    "@cosmiconfig/base": "9.0.0",
-    "@cosmiconfig/types": "9.0.0",
-    "@cosmiconfig/typescript-loader": "9.0.0",
-    "@cosmiconfig/yaml-loader": "9.0.0",
-    "import-fresh": "^3.3.0",
-    "parse-json": "^5.2.0"
-  },
-  "devDependencies": {
-    "@types/parse-json": "^4.0.0",
-    "env-paths": "^2.2.1"
   },
   "engines": {
     "node": ">=14"
