+++ packages/typescript-loader/tsconfig.build.json
@@ -3,14 +3,11 @@
   "compilerOptions": {
     "composite": true,
     "rootDir": "./src",
-    "outDir": "./dist",
-    "stripInternal": true
+    "outDir": "./dist"
   },
   "include": ["src"],
   "references": [
-    { "path": "../base/tsconfig.build.json" },
     { "path": "../types/tsconfig.build.json" },
-    { "path": "../typescript-loader/tsconfig.build.json" },
-    { "path": "../yaml-loader/tsconfig.build.json" }
+    { "path": "../base/tsconfig.build.json" }
   ]
 }
